chore(client): refactor code to use function - change login and signup class to function - exclude client code from webpack - convert dashoard class to addEntry and use function [Deliver #159757576]
