+++ UI/client/addEntry.js
@@ -1,12 +1,9 @@
 const dispalyArea = document.querySelector('.l-display__area');
 const addEntryBtn = document.getElementById('add__Entry');
-const storyTitle = document.querySelector('.story__title');
-const storyText = document.querySelector('.story__text');
-
+const storyTitle = document.querySelector('.story__title')
+const storyText = document.querySelector('.story__text')
 
 const token = JSON.parse(localStorage.getItem('token'));
-addEntryBtn.addEventListener('click', addEntry)
-
 
 async function getEntryOnLoad() {
 
@@ -18,21 +15,18 @@ async function getEntryOnLoad() {
         'Content-type': 'application/json',
       },
     };
-    const res = await fetch(
-      `http://localhost:9000/api/v1/entries`,
-      fetchData
-    );
+    const res = await fetch('http://localhost:9000/api/v1/entries', fetchData);
     const data = await res.json();
     const entries = data.entries;
-    this.getMessage(entries);
+    getMessage(entries);
   }
   if (!token) {
     window.location.href = `${location.protocol}/UI/pages/login.html`;
+
   }
+};
 
-}
-
-function getMessage(entries) {
+const getMessage = (entries) => {
   if (entries.length === 0) {
     const h1 = document.createElement('h1');
     h1.innerHTML = 'You do not have an entry, Please add an entry';
@@ -40,44 +34,17 @@ function getMessage(entries) {
     dispalyArea.appendChild(h1);
   }
   const entry = entries[entries.length - 1];
-  this.displayEntry(entry);
-}
-
-
-function modifyPost(e) {
-  if (e.target.classList.contains('button')) {
-
-  }
-}
-
-async function addEntry() {
-  const title = storyTitle.value;
-  const story = storyText.value;
-  const userData = {
-    title,
-    story,
-  }
-  const fetchData = {
-    method: 'POST',
-    headers: {
-      'Authorization': `Bearer: ${token}`,
-      'Content-type': 'application/json',
-    },
-    body: JSON.stringify(userData)
-  }
-
-  const res = await fetch(
-    `http://localhost:9000/api/v1/entries`,
-    fetchData
-  );
-  const data = await res.json();
-  const entry = data.entry;
   displayEntry(entry);
-}
+};
+
+const modifyPost = (e) => {
+  if (e.target.classList.contains('button')) {
+    console.log('hey edith me');
+  }
+};
 
 
-function displayEntry(entry) {
-  dispalyArea.innerHTML = '';
+const displayEntry = (entry) => {
   // Dairy Title
   const title = document.createElement('span');
   title.className = 'title__display';
@@ -96,7 +63,7 @@ function displayEntry(entry) {
   // Modify Button
   const modifyBtn = document.createElement('button');
   modifyBtn.className = 'edit__post button';
-  modifyBtn.innerHTML = 'Edit';
+  modifyBtn.innerHTML = 'Edith';
 
   // Options
   const option = document.createElement('div');
@@ -107,8 +74,34 @@ function displayEntry(entry) {
   dispalyArea.appendChild(title);
   dispalyArea.appendChild(story);
   dispalyArea.appendChild(option);
-  dispalyArea.addEventListener('click', this.modifyPost);
+  dispalyArea.addEventListener('click', modifyPost);
 }
 
+async function addEntry() {
+  const title = storyTitle.value;
+  const story = storyText.value;
+  const userData = {
+    title,
+    story,
+  };
+  const fetchData = {
+    method: 'POST',
+    headers: {
+      'Authorization': `Bearer: ${token}`,
+      'Content-type': 'application/json',
+    },
+    body: JSON.stringify(userData)
+  };
+  const res = await fetch('http://localhost:9000/api/v1/entries', fetchData);
+  const data = await res.json();
+  if (data.status === 'success') {
+    const entry = data.entry;
+    dispalyArea.innerHTML = '';
+    displayEntry(entry);
+  }
+  console.log(data);
 
-document.addEventListener('DOMContentLoaded', (e) => getEntryOnLoad());
+}
+
+document.addEventListener('DOMContentLoaded', getEntryOnLoad);
+addEntryBtn.addEventListener('click', addEntry);
