--- UI/client/app.js
@@ -1,5 +1,0 @@
-import * as authPage from './pages/AuthPage';
-
-export {
-  authPage,
-}
