--- UI/client/components/Entries/AddEntry.js
@@ -1,21 +1,0 @@
-const addEntryBtn = document.getElementById('add__Entry');
-
-class AddEntry {
-  constructor() {}
-
-
-  addPost() {
-    console.log('add me');
-  }
-
-}
-
-// const addEntry = () => {
-//   console.log('hy')
-// }
-
-// module.export = {
-//   addEntry
-// }
-
-export default new AddEntry;
