--- UI/client/js/Dashboad.js
@@ -1,123 +1,0 @@
-const dispalyArea = document.querySelector('.l-display__area');
-const addEntryBtn = document.getElementById('add__Entry');
-const storyTitle = document.querySelector('.story__title').value;
-const storyText = document.querySelector('.story__text').value;
-
-class Dashboard {
-  constructor() {
-    this.init();
-    this.token;
-  }
-
-  init() {
-    this.token = JSON.parse(localStorage.getItem('token'));
-    this.getEntryOnLoad();
-    addEntryBtn.addEventListener('click', this.addEntry)
-  }
-
-  async getEntryOnLoad() {
-
-    if (this.token) {
-      const fetchData = {
-        method: 'GET',
-        headers: {
-          'Authorization': `Bearer: ${this.token}`,
-          'Content-type': 'application/json',
-        },
-      };
-      const res = await fetch(
-        `http://localhost:9000/api/v1/entries`,
-        fetchData
-      );
-      const data = await res.json();
-      const entries = data.entries;
-      this.getMessage(entries);
-    }
-    if (!this.token) {
-      window.location.href = `${location.protocol}/UI/pages/login.html`;
-    }
-
-  }
-
-  getMessage(entries) {
-    if (entries.length === 0) {
-      const h1 = document.createElement('h1');
-      h1.innerHTML = 'You do not have an entry, Please add an entry';
-      h1.className = 'h-color__white'
-      dispalyArea.appendChild(h1);
-    }
-    const entry = entries[entries.length - 1];
-    this.displayEntry(entry);
-  }
-
-
-  modifyPost(e) {
-    if (e.target.classList.contains('button')) {
-
-      console.log('hey edith me')
-    }
-  }
-
-  async addEntry() {
-    // this.token = JSON.parse(localStorage.getItem('token'));
-    // const userData = {
-    //   title: storyTitle,
-    //   story: storyText
-    // }
-    // const fetchData = {
-    //   method: 'POST',
-    //   headers: {
-    //     'Authorization': `Bearer: ${this.token}`,
-    //     'Content-type': 'application/json',
-    //   },
-    //   body: JSON.stringify(userData)
-    // }
-    // console.log(fetchData)
-    // const res = await fetch(
-    //   `http://localhost:9000/api/v1/entries`,
-    //   fetchData
-    // );
-    // const data = await res.json();
-    // const entry = data.enty;
-
-    // this.displayEntry(entry);
-  }
-
-
-  displayEntry(entry) {
-    // Dairy Title
-    const title = document.createElement('span');
-    title.className = 'title__display';
-    title.innerHTML = `${entry.title}`;
-
-    // Dairy story
-    const story = document.createElement('div');
-    story.className = 'story__display';
-    story.innerHTML = `${entry.story}`;
-
-    // Date created
-    const date = document.createElement('div');
-    date.className = 'date__posted';
-    date.innerHTML = `${entry.created_at}`;
-
-    // Modify Button
-    const modifyBtn = document.createElement('button');
-    modifyBtn.className = 'edit__post button';
-    modifyBtn.innerHTML = 'Edith';
-
-    // Options
-    const option = document.createElement('div');
-    option.className = 'l-story__option';
-    option.appendChild(date);
-    option.appendChild(modifyBtn);
-
-    dispalyArea.appendChild(title);
-    dispalyArea.appendChild(story);
-    dispalyArea.appendChild(option);
-    dispalyArea.addEventListener('click', this.modifyPost);
-  }
-
-}
-
-
-document.addEventListener('DOMContentLoaded', (e) => new Dashboard());
