+++ UI/pages/dashboard.html
@@ -112,8 +112,7 @@
       </div>
     </section>
     <!-- End Main Content -->
-    <script src="../build/bundle.js"></script>
-    <script src="../client/js/AddEntry.js"></script>
+    <script src="../client/addEntry.js"></script>
 </body>
 
 </html>
