+++ UI/pages/login.html
@@ -68,6 +68,6 @@
   </div>
   </div>
 </body>
-<script src="../build/bundle.js"></script>
+<script src="../client/login.js"></script>
 
 </html>
