+++ UI/pages/signup.html
@@ -72,6 +72,6 @@
   </div>
   </div>
 </body>
-<script src="../build/bundle.js"></script>
+<script src="../client/signup.js"></script>
 
 </html>
